fix(UserAccess): Moving query endpoint to access groups service
